Update main.py
added the add_task  in order to be able to add a task while also organizing them by date. If the date is not to be found it will create a new entry with an empty list. If the date does in fact already exist it will append the task to the already existing list for the specific date.