Update main.py
added display_task to checks if the date argument exists as a key . If the date appears it means that there are tasks affiliated with the date.However, if the date is not found it means there are no tasks for the specific date. If tasks are found for the date provided it will print a message indicating the date and 
 proceeds to repeat over the tasks using a loop.
Inside the loop, it  will print each task with a corresponding number which will start at 1 using the enumerate() function.
As well, if there are no tasks for the given date, it prints a message that says no tasks were found.