Update main.py
added def_display this calls the monthdaycalendar function from the calendar. Additionally, created a head string in order to link the name of each month with the year. The addition of header.center prints the header within 20 characters.Furthermore printing the names of the days of the week. If the day is 0, it means that it is outside the month, and it prints empty spaces hence the usage of ("").
If the day is not 0, it will check if the string of the day is found in  the list called self.tasks. If it is , it assigns tasks_indicator a value of "*", indicating that there are tasks for that day. If not it assigns  " " a space.